backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opensearch-py==2.6.0
python-dotenv==1.0.1
pydantic==2.9.2